String work
- Improved CString and String handling
- Improved argument and return type handling in general
- Fixed codegen.py get_size function not honoring the context the struct was declared in
- Fixed a bug in builtin detection
- Fixed a bug in sizeof type matching
- Added possibility to specify stdlib imports without prefixing them with 'rial:'
- Added deduplication to string literal parsing
- Fixed string literal size not being recorded
- Fixed DesugarTransformer.py using old Transformer
- Fixed implicit parameter being appended to the end of the list of arguments instead of the front
- Fixed array parsing for function calls
- Changed default optimization level to '1' to improve default performance
- Fixed CString (in particular) handling by return and variable assignment rules
- Fixed variable detection in RIALVariable.py and added __str__ and __repr__ functions

+++ rial/transformer/builtin_type_to_llvm_mapper.py
@@ -21,7 +21,7 @@ def null(ty):
 @lru_cache(24)
 def is_builtin_type(ty: str):
     ty = map_shortcut_to_type(ty)
-    return ty in ("Int32", "Int64", "UInt64", "UInt64", "Double64", "Float32", "Boolean", "Byte", "Char", "Half")
+    return ty in ("Int32", "Int64", "UInt32", "UInt64", "Double64", "Float32", "Boolean", "Byte", "Char", "Half")
 
 
 @lru_cache(24)
